% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_usb.R
\name{accel_fix_usb}
\alias{accel_fix_usb}
\title{Address breaks in the primary accelerometer data stream due to device USB
connection}
\usage{
accel_fix_usb(records, info)
}
\arguments{
\item{records}{list of parsed packets}

\item{info}{result of \code{\link{parse_info_txt}}}
}
\description{
Address breaks in the primary accelerometer data stream due to device USB
connection
}
\keyword{internal}
