% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latch_accelerometer_records.R
\name{accel_zero_fill}
\alias{accel_zero_fill}
\alias{accel_zero_fill.list}
\alias{accel_zero_fill.data.frame}
\title{Fill in gaps in accelerometer data with zeroes}
\usage{
accel_zero_fill(records, index, n_missing_rows, info, ...)

\method{accel_zero_fill}{list}(records, index, n_missing_rows, info, ...)

\method{accel_zero_fill}{data.frame}(records, index, n_missing_rows, info, ...)
}
\arguments{
\item{records}{the records to be examined and filled in}

\item{n_missing_rows}{The number of zero values to fill in}

\item{info}{result of \code{\link{parse_info_txt}}}
}
\description{
Fill in gaps in accelerometer data with zeroes
}
\keyword{internal}
