% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{activity_payload}
\alias{activity_payload}
\title{Parse the payload for an ACTIVITY packet}
\usage{
activity_payload(payload, samp_rate, scale_factor, is_last_packet)
}
\description{
Parse the payload for an ACTIVITY packet
}
\keyword{internal}
