% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_3x_read.R
\name{AG_binary}
\alias{AG_binary}
\title{Convert a raw elements to a binary sequence}
\usage{
AG_binary(value, n = 8)
}
\arguments{
\item{value}{raw. The elements to convert}

\item{n}{integer. Desired length of output for each element}
}
\description{
Essentially a wrapper for \code{\link[binaryLogic]{as.binary}}
}
\keyword{internal}
