% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Counts_Utils.R
\name{AG_col_names}
\alias{AG_col_names}
\title{Map variable inclusion to columns in csv file}
\usage{
AG_col_names(variable)
}
\arguments{
\item{variable}{A variable included in the output}
}
\description{
Map variable inclusion to columns in csv file
}
\keyword{internal}
