% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Raw_Utils.R
\name{AG_collapse}
\alias{AG_collapse}
\title{Collapse primary accelerometer data}
\usage{
AG_collapse(AG, output_window_secs = 1, samp_freq, method = "default",
  ENMO2 = NULL)
}
\arguments{
\item{AG}{a dataframe of raw primary accelerometer data}

\item{output_window_secs}{the desired epoch length; defaults to one second}

\item{samp_freq}{The sampling frequency}

\item{method}{character scalar giving the method to use for calculating ENMO,
either "default" or "block"}

\item{ENMO2}{vector of leftover raw values from the previous block (if
applicable)}
}
\description{
Collapse primary accelerometer data
}
\examples{
data(raw_to_collapse)
collapsed <- AG_collapse(raw_to_collapse, 1, 80)
head(collapsed)

}
