% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_IMU_Format.R
\name{ag_imu_format}
\alias{ag_imu_format}
\title{Format columns in collapsed IMU data}
\usage{
ag_imu_format(AG, output_window_secs = 1, filter, samp_rate, filter_hz = 35,
  verbose, block_size, return_raw = FALSE)
}
\arguments{
\item{AG}{data frame containing IMU data}

\item{output_window_secs}{the desired epoch length; defaults to one second}

\item{filter}{a logical scalar: Apply a low-pass filter to gyroscope data?}

\item{samp_rate}{integer. The sampling rate.}

\item{filter_hz}{The cutoff for the low-pass filter}

\item{verbose}{A logical scalar: Print processing updates?}

\item{block_size}{integer. Rows of raw data per output window}

\item{return_raw}{logical. Return raw triaxial data?}
}
\description{
Format columns in collapsed IMU data
}
\keyword{internal}
