% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Raw_Utils.R
\name{ag_raw_format}
\alias{ag_raw_format}
\title{Format columns in collapsed raw data}
\usage{
ag_raw_format(AG, start_time, output_window_secs = 1)
}
\arguments{
\item{AG}{data frame containing raw data}

\item{start_time}{the start time for calculating timestamps}

\item{output_window_secs}{the desired epoch length; defaults to one second}
}
\description{
Format columns in collapsed raw data
}
\keyword{internal}
