% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_utils.R
\name{AG_round}
\alias{AG_round}
\title{Function to round away from zero}
\usage{
AG_round(x, digits = 6)
}
\arguments{
\item{x}{the value to round}

\item{digits}{digits to round to}
}
\description{
Function to round away from zero
}
\keyword{internal}
