% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Counts.R
\name{AG_time}
\alias{AG_time}
\title{Add time variable to processed data}
\usage{
AG_time(AG, meta)
}
\arguments{
\item{AG}{A data frame of partially-processed data}

\item{meta}{A list given by \code{\link{AG_meta}}}
}
\description{
Add time variable to processed data
}
\keyword{internal}

