% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGread.R
\docType{package}
\name{AGread}
\alias{AGread}
\alias{AGread-package}
\title{Read Accelerometer Files from ActiGraph Accelerometers}
\description{
This provides support for reading ActiGraph files of various modes into R.
For more information see:
\url{https://actigraph.desk.com/customer/en/portal/articles/2515800-what-do-the-different-mode-numbers-mean-in-a-csv-or-dat-file-}
}
\section{Core function}{

\code{\link{read_AG_counts}}
}

\examples{
\dontrun{
read_AG_counts(system.file("extdata", "Example.csv", package = "AGread"))
}

}
