% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_3x_read.R
\name{bin_int}
\alias{bin_int}
\title{Calculate an integer from a binary sequence}
\usage{
bin_int(value_bin)
}
\arguments{
\item{value_bin}{The binary sequence}
}
\description{
Calculate an integer from a binary sequence
}
\keyword{internal}
