% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_raw.R
\name{calibrate_raw}
\alias{calibrate_raw}
\title{Perform autocalibration using GGIR package}
\usage{
calibrate_raw(AG, file)
}
\arguments{
\item{AG}{A dataframe of raw acceleration values}

\item{file}{A character scalar giving path to primary accelerometer file}
}
\value{
A dataframe of calibrated acceleration values
}
\description{
Perform autocalibration using GGIR package
}
\keyword{internal}
