% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_utils.R
\name{capability_collapse}
\alias{capability_collapse}
\title{Collapse a character vector for tidy printing}
\usage{
capability_collapse(caps)
}
\arguments{
\item{caps}{the vector to collapse}
}
\description{
Collapse a character vector for tidy printing
}
\keyword{internal}
