% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Counts_Utils.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize a string vector}
\usage{
capitalize(string)
}
\arguments{
\item{string}{the vector to capitalize. Scalars accepted.}
}
\description{
First character will be capitalized, and all others remain unchanged
}
\keyword{internal}
