% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_id}
\alias{check_id}
\title{Check sensor payload ID prior to parsing the packet}
\usage{
check_id(x, payload)
}
\arguments{
\item{x}{the raw payload}

\item{payload}{the sensor schema payload (as list) to compare against}
}
\description{
Check sensor payload ID prior to parsing the packet
}
\keyword{internal}
