% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Function.R
\name{check_inc}
\alias{check_inc}
\title{Test whether inclinometer variables have been correctly assigned}
\usage{
check_inc(AG)
}
\arguments{
\item{AG}{A data frame of partially-processed data}
}
\description{
Test whether inclinometer variables have been correctly assigned
}
\keyword{internal}
