% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_utils.R
\name{check_packets}
\alias{check_packets}
\title{Check for packets with failed checksum}
\usage{
check_packets(packets)
}
\arguments{
\item{packets}{object containing data packets or subsequently parsed data}
}
\description{
Check for packets with failed checksum
}
\keyword{internal}
