% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_Reading_Helper_Functions.R
\name{check_second}
\alias{check_second}
\title{Check if the IMU data start on an exact second}
\usage{
check_second(AG)
}
\arguments{
\item{AG}{a dataframe of IMU data}
}
\description{
Check if the IMU data start on an exact second
}
\examples{
data(imu_to_check)
check_second(imu_to_check)

}

