% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checksum_calculate.R
\name{checksum_calculate}
\alias{checksum_calculate}
\title{Perform a checksum calculation on gt3x file packets}
\usage{
checksum_calculate(record, final_index)
}
\arguments{
\item{record}{raw. Packet from a gt3x file}

\item{final_index}{integer. Index of the final packet byte}
}
\description{
Perform a checksum calculation on gt3x file packets
}
\keyword{internal}
