% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{checksumC}
\alias{checksumC}
\title{Calculate checksum for a packet in C++}
\usage{
checksumC(log, start_index, end_index)
}
\arguments{
\item{log}{RawVector representing the contents of log.bin}

\item{start_index}{the packet start index}

\item{end_index}{the packet end index}
}
\description{
Calculate checksum for a packet in C++
}
\keyword{internal}
