% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_packet_set.R
\name{collapse_packet_set}
\alias{collapse_packet_set}
\alias{collapse_packet_set.default}
\title{Collapse a set of processed gt3x packets}
\usage{
collapse_packet_set(set, results, ...)

\method{collapse_packet_set}{default}(set, results, ...)
}
\arguments{
\item{set}{the set of record headers corresponding to each packet}

\item{results}{the processed packet information}

\item{...}{further arguments passed to methods}
}
\description{
Collapse a set of processed gt3x packets
}
\keyword{internal}
