% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_utils.R
\name{cpp_include}
\alias{cpp_include}
\title{Switch character packet inclusion vector to an integer vector for Rcpp dev
parser}
\usage{
cpp_include(include)
}
\arguments{
\item{include}{character. The PACKET types to parse}
}
\description{
Switch character packet inclusion vector to an integer vector for Rcpp dev
parser
}
\keyword{internal}
