% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_bin_dev_packets.R
\name{get_parameters}
\alias{get_parameters}
\alias{get_schema}
\alias{get_events}
\alias{get_activity2}
\title{Parse gt3x packets using the dev scheme}
\usage{
get_parameters(packets, tz, verbose)

get_schema(packets, tz, verbose)

get_events(packets, tz, info, verbose)

get_activity2(packets, all_times, tz, info, verbose)
}
\arguments{
\item{packets}{list object containing data packets}

\item{tz}{character. The timezone to use}

\item{verbose}{logical. Print updates to console?}

\item{all_times}{vector of all expected timestamps}

\item{packets}{list object containing data packets}

\item{packets}{list object containing data packets}
}
\description{
Parse gt3x packets using the dev scheme
}
\keyword{internal}
