% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_utils.R
\docType{data}
\name{.accel_names}
\alias{.accel_names}
\title{Variable names for RAW objects}
\format{An object of class \code{character} of length 3.}
\usage{
.accel_names
}
\description{
Variable names for RAW objects
}
\keyword{internal}
