% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_utils.R
\name{empty_raw}
\alias{empty_raw}
\title{Construct missing packet entries for ACTIVITY2 (RAW) data}
\usage{
empty_raw(timestamps, empty_value, info)
}
\arguments{
\item{timestamps}{the packet timestamps}

\item{empty_value}{the value to assign missing accelerometer entries}

\item{info}{output of \code{\link{parse_info_txt}}}
}
\description{
Construct missing packet entries for ACTIVITY2 (RAW) data
}
\keyword{internal}
