% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_usb.R
\name{fix_usb}
\alias{fix_usb}
\alias{fix_usb.SENSOR_DATA}
\title{Repair packet streams interrupted by USB connection events}
\usage{
fix_usb(object, ...)

\method{fix_usb}{SENSOR_DATA}(records)
}
\arguments{
\item{object}{the sensor stream to repair}

\item{...}{further arguments to methods}
}
\description{
Repair packet streams interrupted by USB connection events
}
\keyword{internal}
