% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lux.R
\name{format_lux}
\alias{format_lux}
\title{Convert pre-parsed lux data into its final format}
\usage{
format_lux(lux, info)
}
\arguments{
\item{lux}{output from \code{dev_parse_lux}, with minor modifications}

\item{info}{data frame containing the lux factors}
}
\description{
Convert pre-parsed lux data into its final format
}
\keyword{internal}
