% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sleep_events.R
\name{format_sleep_events}
\alias{format_sleep_events}
\title{Format idle sleep information from EVENT records}
\usage{
format_sleep_events(sleeps, tz, verbose)
}
\arguments{
\item{sleeps}{data frame with idle sleep mode information}

\item{tz}{character. The timezone to use}

\item{verbose}{logical. Print updates to console?}
}
\description{
Format idle sleep information from EVENT records
}
\keyword{internal}
