% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reintegrate_Utils.R
\name{test_second}
\alias{test_second}
\alias{forwards_start}
\alias{backwards_start}
\alias{get_blocks}
\title{Assign blocks to a data stream for reintegration}
\usage{
test_second(timestamp, to)

forwards_start(ag, time_var, to, block_size, begin = 1)

backwards_start(ag, time_var, to, block_size, begin = 1)

get_blocks(ag, time_var, to, start_epoch, direction)
}
\arguments{
\item{timestamp}{POSIX scalar}

\item{to}{the desired epoch length of the output. Starting epoch length will
be determined automatically (while also verifying the data are continuous).}

\item{ag}{a data frame to reintegrate}

\item{time_var}{The name of the column containing POSIX-formatted timestamp
information}

\item{block_size}{The number of rows of data included in each reintegrated
epoch}

\item{begin}{A starting index}

\item{start_epoch}{The initial epoch length of the data being reintegrated}

\item{direction}{The direction of reintegration, i.e. whether a timestamp
refers to the timespan after the previous data point ("backwards"), or
before the next data point ("forwards").}
}
\description{
Assign blocks to a data stream for reintegration
}
\keyword{internal}
