% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates&Times.R
\name{get_day_of_year}
\alias{get_day_of_year}
\title{Julian Date}
\usage{
get_day_of_year(timestamp, format = "\%Y-\%m-\%d \%H:\%M:\%S")
}
\arguments{
\item{timestamp}{A character vector containing timestamp information}

\item{format}{The date-time format of the \code{timestamp} vector}
}
\value{
A numeric vector of Julian dates.
}
\description{
A wrapper to retrieve the Julian date.
}
\examples{
\dontrun{
key_dates <- c("2018-01-01", "2018-12-31")

AGread:::get_day_of_year(key_dates, "\%Y-\%m-\%d")
}

}
\keyword{internal}
