% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0d_Deprecated_Doc.R
\name{get_duration-deprecated}
\alias{get_duration-deprecated}
\title{Provide the run time of processing}
\usage{
get_duration(timer)
}
\arguments{
\item{timer}{a proc_time object giving the initial time}
}
\description{
Provide the run time of processing
}
\examples{
timer <- proc.time()
Sys.sleep(2.2)
\donttest{
  # Gives warning about deprecation
  # get_duration(timer)

  # Instead use
  PAutilities::get_duration(timer)
}


}
\seealso{
\code{\link{AGread-deprecated}}
}
\keyword{internal}
