% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Counts.R
\name{get_duration}
\alias{get_duration}
\title{Provide the run time of processing}
\usage{
get_duration(timer)
}
\arguments{
\item{timer}{The initial time}
}
\description{
Provide the run time of processing
}
\examples{
timer <- proc.time()
Sys.sleep(2.2)
AGread:::message_update(16, dur = get_duration(timer))

}
