% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGread_utils.R
\name{get_epoch}
\alias{get_epoch}
\title{Retrieve the epoch length of an ActiGraph data frame}
\usage{
get_epoch(AG, time_var = "Timestamp")
}
\arguments{
\item{AG}{data frame of ActiGraph data}

\item{time_var}{character scalar. Name of the time variable column}
}
\value{
scalar epoch length
}
\description{
Retrieve the epoch length of an ActiGraph data frame
}
\examples{
AG <- read_AG_counts(
  system.file("extdata/example1sec.csv", package = "AGread"),
  header = TRUE
)
get_epoch(AG)
}
