% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_size.R
\name{get_file_size__gb}
\alias{get_file_size__gb}
\title{Get the size of a file in GB}
\usage{
get_file_size__gb(file)
}
\arguments{
\item{file}{A character scalar giving path to primary accelerometer file}
}
\description{
Get the size of a file in GB
}
\keyword{internal}
