% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_utils.R
\name{get_float_value}
\alias{get_float_value}
\alias{get_exponent}
\alias{get_significand}
\title{Retrieve the float value for appropriate PARAMETERS records}
\usage{
get_float_value(value)

get_exponent(value, n_bytes)

get_significand(value, n_bits, n_bytes)
}
\arguments{
\item{value}{raw. The record to convert}
}
\description{
Retrieve the float value for appropriate PARAMETERS records
}
\keyword{internal}
