% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_record_header.R
\name{get_headers}
\alias{get_headers}
\title{Retrieve record headers from .gt3x binary data}
\usage{
get_headers(log, n_records, tz = "UTC", verbose = FALSE)
}
\arguments{
\item{log}{raw. The data from log.bin}

\item{n_records}{integer. The length of log.bin}

\item{tz}{character. The timezone to use}

\item{verbose}{logical. Print updates to console?}
}
\description{
Retrieve record headers from .gt3x binary data
}
\keyword{internal}
