% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_int12}
\alias{get_int12}
\title{Extract a 12-bit integer from two bytes}
\usage{
get_int12(x, i1, i2, full_first)
}
\arguments{
\item{x}{the bytes (RawVector) from which to extract the integer}

\item{i1}{integer. The index of the first byte}

\item{i2}{integer. The index of the second byte}

\item{full_first}{boolean. Use the full first byte and half of the second?}
}
\description{
Extract a 12-bit integer from two bytes
}
\keyword{internal}
