% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_SENSOR_DATA.R
\name{get_sensor_column}
\alias{get_sensor_column}
\title{Parse a column of sensor data}
\usage{
get_sensor_column(
  payload,
  indices,
  bytesInValue,
  is_signed,
  endianness,
  scale_factor,
  label
)
}
\arguments{
\item{payload}{raw. The payload to parse}

\item{indices}{vector giving the starting index of each record to be parsed}

\item{bytesInValue}{integer. Length of payload entry}

\item{is_signed}{logical. Is encoded value signed?}

\item{endianness}{character. Is endianness big or little?}

\item{scale_factor}{numeric. Scale factor for binary-to-unit conversions}

\item{label}{character. The column label}
}
\description{
Parse a column of sensor data
}
\keyword{internal}
