% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SENSOR_DATA_payload.R
\name{get_sensor_row}
\alias{get_sensor_row}
\title{Parse a row of sensor data}
\usage{
get_sensor_row(payload, column, offset, BITS_PER_BYTE)
}
\arguments{
\item{payload}{raw. The payload to parse}

\item{column}{data frame. Information about the column in which the result
will be placed}

\item{offset}{integer. The starting index for parsing}

\item{BITS_PER_BYTE}{integer. The number of bits per byte}
}
\description{
Parse a row of sensor data
}
\keyword{internal}
