% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_short}
\alias{get_short}
\title{Flexibly (big/little endian, signed/unsigned) convert two raw bytes to short}
\usage{
get_short(x, i1, i2, is_signed)
}
\arguments{
\item{x}{the bytes (RawVector) from which to extract the short}

\item{i1}{integer. The index of the first byte}

\item{i2}{integer. The index of the second byte}

\item{is_signed}{boolean. Return a signed value?}
}
\description{
Flexibly (big/little endian, signed/unsigned) convert two raw bytes to short
}
\keyword{internal}
