% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_times.R
\name{get_actual}
\alias{get_actual}
\alias{get_expected}
\alias{get_seq}
\title{Set up timestamps for RAW or IMU data frames}
\usage{
get_actual(packets, tz)

get_expected(start, end, samp_rate)

get_seq(start, end, samp_rate)
}
\arguments{
\item{packets}{object containing data packets or subsequently parsed data}

\item{tz}{character. The timezone to use}

\item{start}{start time}

\item{end}{end time}

\item{samp_rate}{sampling rate}
}
\description{
Set up timestamps for RAW or IMU data frames
}
\keyword{internal}
