% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_3x_read.R
\name{get_value}
\alias{get_value}
\title{Evaluate the value portion of a PARAMETERS record}
\usage{
get_value(type, value)
}
\arguments{
\item{type}{character. The record type}

\item{value}{raw. The record to evaluate}
}
\description{
Evaluate the value portion of a PARAMETERS record
}
\keyword{internal}
