% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_IMU.R
\name{impute_IMU}
\alias{impute_IMU}
\title{Impute IMU data for missing packets}
\usage{
impute_IMU(object, verbose)
}
\arguments{
\item{object}{data frame of IMU data}

\item{verbose}{logical. Print updates to console?}
}
\description{
Impute IMU data for missing packets
}
\keyword{internal}
