% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imu_df}
\alias{imu_df}
\title{Convert parsed packet data from list to data frame}
\usage{
imu_df(input)
}
\arguments{
\item{input}{parsed packet data (as list)}
}
\description{
Convert parsed packet data from list to data frame
}
\keyword{internal}
