% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMU_Filter.R
\name{imu_filter_gyroscope}
\alias{imu_filter_gyroscope}
\title{Low-Pass filter Gyroscope data}
\usage{
imu_filter_gyroscope(AG, samp_rate, filter_hz = 35, verbose = FALSE)
}
\arguments{
\item{AG}{a dataframe of IMU data}
}
\description{
Low-Pass filter Gyroscope data
}
\examples{
\dontrun{
data(imu_to_collapse)
AGread:::imu_filter_gyroscope(imu)
}

}
\keyword{internal}
