% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_usb.R
\name{imu_fix_usb}
\alias{imu_fix_usb}
\title{Address breaks in the IMU data stream due to device USB connection}
\usage{
imu_fix_usb(records)
}
\arguments{
\item{records}{list of parsed packets}
}
\description{
Address breaks in the IMU data stream due to device USB connection
}
\keyword{internal}
