% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_IMU_Utils.R
\name{imu_name_label}
\alias{imu_name_label}
\title{Fix the column names in IMU data during collapsing}
\usage{
imu_name_label(pattern, replacement, label, AG_names)
}
\arguments{
\item{pattern}{string pattern to replace}

\item{replacement}{replacement pattern}

\item{label}{additional label for the replacement (i.e., units)}

\item{AG_names}{column names of the IMU data}
}
\description{
Fix the column names in IMU data during collapsing
}
\keyword{internal}
