% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_IMU_Utils.R
\name{imu_var_collapse}
\alias{imu_var_collapse}
\title{Collapse an IMU variable}
\usage{
imu_var_collapse(AG, var, absolute = FALSE, label = "")
}
\arguments{
\item{AG}{the IMU data frame to select from}

\item{var}{the character pattern to use in determining which columns to
collapse}

\item{absolute}{logical. Should columns be converted to absolute values?}

\item{label}{additional information (i.e., units) to include in variable
names}
}
\description{
Collapse an IMU variable
}
\keyword{internal}
