% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interval_match}
\alias{interval_match}
\title{Match a resampled interval proportion to its corresponding
originally-sampled interval in C++.}
\usage{
interval_match(proportion, references)
}
\arguments{
\item{proportion}{double. The interval proportion to match.}

\item{reference_frame}{DataFrame containing a \code{prop_min} column to
serve as originally-sampled interval proportions}
}
\description{
Match a resampled interval proportion to its corresponding
originally-sampled interval in C++.
}
\keyword{internal}
