% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latch_accelerometer_records.R
\name{latch_accelerometer_records}
\alias{latch_accelerometer_records}
\title{Fill in accelerometer values for periods when the device was connected to USB}
\usage{
latch_accelerometer_records(missing_records, records, info)
}
\arguments{
\item{missing_records}{logical vector indicating whether a record is labeled
as missing}

\item{records}{the records to be examined and filled in}

\item{info}{result of \code{\link{parse_info_txt}}}
}
\description{
Fill in accelerometer values for periods when the device was connected to USB
}
\keyword{internal}
