% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latch_gt3x.R
\name{latch_gt3x}
\alias{latch_gt3x}
\alias{check_missing}
\title{Replace imputed zeroes by carrying forward last non-zero value}
\usage{
latch_gt3x(
  AG,
  x_var = "Accelerometer_X",
  y_var = "Accelerometer_Y",
  z_var = "Accelerometer_Z",
  verbose = FALSE,
  flag_idle_sleep = FALSE,
  ...,
  is_sleep = NULL
)

check_missing(
  AG,
  x_var = "Accelerometer_X",
  y_var = "Accelerometer_Y",
  z_var = "Accelerometer_Z",
  ...
)
}
\arguments{
\item{AG}{data frame of ActiGraph data (could be an \code{activity_df} object
from the \code{read.gt3x} package)}

\item{x_var, y_var, z_var}{character scalars naming the three columns of raw
acceleration data in \code{AG}}

\item{verbose}{logical. Print updates to console?}

\item{flag_idle_sleep}{should recorded idle sleep times be tagged?}

\item{...}{arguments passed to \code{check_missing}}

\item{is_sleep}{an optional logical vector whose length must equal
\code{nrow(AG)}. Can be passed as an override to internal calculation of
idle sleep via \code{check_missing}}

\item{time_var}{character scalar naming the column of \code{AG} containing
POSIX-formatted timestamps}
}
\value{
An updated data frame with latched values rather than zeroes for
  missing data
}
\description{
Replace imputed zeroes by carrying forward last non-zero value
}
\note{
\code{check_missing} may pick up on more than idle sleep mode (e.g.,
  USB connect events or trailing zeroes at the end of a file). To exclusively
  isolate idle sleep mode, you will need to go through some of the other
  \code{AGread::read_gt3x} avenues that include event parsing.
}
\examples{
\donttest{
file_3x <- system.file(
  "extdata", "example.gt3x", package = "AGread"
)
AG <- read_gt3x(file_3x)$RAW
head(latch_gt3x(AG))
}

}
\keyword{internal}
