% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{latch_packet}
\alias{latch_packet}
\title{Fill in a packet by latching to the end of the previous packet}
\usage{
latch_packet(last_packet, zero_packet, sample_rate)
}
\arguments{
\item{last_packet}{the previous packet}

\item{zero_packet}{a zero_valued packet}

\item{sample_rate}{the sampling rate}
}
\description{
Fill in a packet by latching to the end of the previous packet
}
\keyword{internal}
