% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latch_value.R
\name{latch_value}
\alias{latch_value}
\title{Latch a data packet to the last payload sample of the previous packet}
\usage{
latch_value(records, index)
}
\description{
Latch a data packet to the last payload sample of the previous packet
}
\keyword{internal}
