% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_legacy_dev_compare.R
\name{legacy_dev_compare}
\alias{legacy_dev_compare}
\title{Verify \code{\link{read_gt3x}} gives equivalent output using legacy and dev
parsers}
\usage{
legacy_dev_compare(legacy, dev)
}
\arguments{
\item{legacy}{legacy output}

\item{dev}{dev output}
}
\value{
A logical scalar indicating whether the outputs matched
}
\description{
Verify \code{\link{read_gt3x}} gives equivalent output using legacy and dev
parsers
}
\examples{
\donttest{
file_3x <- system.file(
  "extdata", "example.gt3x", package = "AGread"
)
l <- read_gt3x(file_3x, parser = "legacy")
d <- read_gt3x(file_3x, parser = "dev")
legacy_dev_compare(l, d)
}
}
