% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{legacy_payload_parse_activity2_26C}
\alias{legacy_payload_parse_activity2_26C}
\alias{dev_activity2_payload}
\title{Parse a packet of primary accelerometer data in C++}
\usage{
legacy_payload_parse_activity2_26C(
  payload,
  samp_rate,
  scale_factor,
  is_last_packet,
  timestamp
)

dev_activity2_payload(payload, samp_rate, scale_factor, is_last_packet)
}
\arguments{
\item{payload}{RawVector containing the payload bytes}

\item{samp_rate}{integer reflecting the sampling rate}

\item{scale_factor}{integer reflecting the scale factor}

\item{is_last_packet}{logical. Is this the last packet in the file?}

\item{timestamp}{Datetime. The packet timestamp}
}
\description{
Parse a packet of primary accelerometer data in C++
}
\keyword{internal}
