% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lux_payload}
\alias{lux_payload}
\title{Parse the payload for a LUX packet}
\usage{
lux_payload(payload)
}
\arguments{
\item{payload}{RawVector. The payload}
}
\description{
Parse the payload for a LUX packet
}
\keyword{internal}
