% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{message_update}
\alias{message_update}
\title{Print processing updates to console}
\usage{
message_update(n, is_message = FALSE, file)
}
\arguments{
\item{n}{The message number to print}

\item{is_message}{A logical scalar: Print as message?}

\item{file}{A character scalar: name of file being processed}
}
\description{
Print processing updates to console
}
\keyword{internal}
