% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mid_round}
\alias{mid_round}
\title{Perform midpoint rounding the ActiGraph way}
\usage{
mid_round(input, digits)
}
\arguments{
\item{input}{double. The number to round}

\item{digits}{int. The number of digits to round to}
}
\description{
This function is inefficient but necessary to ensure alignment between
\code{\link{read_gt3x}} and \code{\link{read_AG_raw}}
}
\keyword{internal}
