% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_log_bin.R
\name{name_log}
\alias{name_log}
\title{Lightly format a processed log}
\usage{
name_log(log)
}
\arguments{
\item{log}{the log to process}
}
\description{
Lightly format a processed log
}
\keyword{internal}
