% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{next_separator}
\alias{next_separator}
\title{Find the next record separator}
\usage{
next_separator(log, index)
}
\arguments{
\item{log}{RawVector. The contents of log.bin}

\item{index}{int. The starting index from which to search for a record
separator}
}
\description{
Find the next record separator
}
\keyword{internal}
