% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{num_pb}
\alias{num_pb}
\title{Combine numeric vectors}
\usage{
num_pb(a, b)
}
\arguments{
\item{a}{the vector to be appended}

\item{b}{the vector with which to append \code{a}}
}
\description{
Combine numeric vectors
}
\keyword{internal}
