% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_log_bin_utils.R
\name{packet_print}
\alias{packet_print}
\title{Print packet-parsing information to console}
\usage{
packet_print(type = c("startup", "progress", "cleanup"), label, i)
}
\arguments{
\item{type}{character. The message type}

\item{label}{character. The packet type}

\item{i}{numeric. Proportion to print as percentage for progress updates}
}
\description{
If \code{verbose} has been set to \code{TRUE} in the parent function, this
function will be invoked to construct the message and print it.
}
\keyword{internal}
