% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMETERS_payload_parsers_PARENT.R
\name{par_battery_state}
\alias{par_battery_state}
\title{Parse PARAMETERS packet for battery state}
\usage{
par_battery_state(value, key)
}
\arguments{
\item{value}{portion of payload to parse}

\item{key}{PARAMETERS key entry to use for guidance}
}
\description{
Parse PARAMETERS packet for battery state
}
\keyword{internal}
