% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMETERS_payload_parsers_CHILD.R,
%   R/PARAMETERS_payload_parsers_PARENT.R
\name{par_board_revision}
\alias{par_board_revision}
\alias{par_imu_accel_scale}
\alias{par_imu_gyro_scale}
\alias{par_imu_mag_scale}
\alias{par_accel_scale}
\alias{par_imu_temp_scale}
\alias{par_imu_temp_offset}
\alias{par_negative_g_offset_x}
\alias{par_negative_g_offset_y}
\alias{par_negative_g_offset_z}
\alias{par_positive_g_offset_x}
\alias{par_positive_g_offset_y}
\alias{par_positive_g_offset_z}
\alias{par_sample_rate}
\alias{par_zero_g_offset_x}
\alias{par_zero_g_offset_y}
\alias{par_zero_g_offset_z}
\alias{par_imu_negative_g_offset_x}
\alias{par_imu_negative_g_offset_y}
\alias{par_imu_negative_g_offset_z}
\alias{par_imu_positive_g_offset_x}
\alias{par_imu_positive_g_offset_y}
\alias{par_imu_positive_g_offset_z}
\alias{par_utc_offset}
\alias{par_imu_zero_g_offset_x}
\alias{par_imu_zero_g_offset_y}
\alias{par_imu_zero_g_offset_z}
\alias{par_battery_voltage}
\title{Parse PARAMETERS packet for battery voltage}
\usage{
par_board_revision(value, key)

par_imu_accel_scale(value, key)

par_imu_gyro_scale(value, key)

par_imu_mag_scale(value, key)

par_accel_scale(value, key)

par_imu_temp_scale(value, key)

par_imu_temp_offset(value, key)

par_negative_g_offset_x(value, key)

par_negative_g_offset_y(value, key)

par_negative_g_offset_z(value, key)

par_positive_g_offset_x(value, key)

par_positive_g_offset_y(value, key)

par_positive_g_offset_z(value, key)

par_sample_rate(value, key)

par_zero_g_offset_x(value, key)

par_zero_g_offset_y(value, key)

par_zero_g_offset_z(value, key)

par_imu_negative_g_offset_x(value, key)

par_imu_negative_g_offset_y(value, key)

par_imu_negative_g_offset_z(value, key)

par_imu_positive_g_offset_x(value, key)

par_imu_positive_g_offset_y(value, key)

par_imu_positive_g_offset_z(value, key)

par_utc_offset(value, key)

par_imu_zero_g_offset_x(value, key)

par_imu_zero_g_offset_y(value, key)

par_imu_zero_g_offset_z(value, key)

par_battery_voltage(value, key)
}
\arguments{
\item{value}{portion of payload to parse}

\item{key}{PARAMETERS key entry to use for guidance}
}
\description{
Parse PARAMETERS packet for battery voltage
}
\keyword{internal}
