% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_PARAMETERS_CHILD.R,
%   R/payload_PARAMETERS_PARENT.R
\name{par_target_start_time}
\alias{par_target_start_time}
\alias{par_target_stop_time}
\alias{par_time_of_day}
\alias{par_calibration_time}
\title{Parse PARAMETERS packet for calibration time}
\usage{
par_target_start_time(value, key, tz)

par_target_stop_time(value, key, tz)

par_time_of_day(value, key, tz)

par_calibration_time(value, key, tz)
}
\arguments{
\item{value}{portion of payload to parse}

\item{key}{PARAMETERS key entry to use for guidance}

\item{tz}{character. Time zone to use.}
}
\description{
Parse PARAMETERS packet for calibration time
}
\keyword{internal}
