% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_PARAMETERS_PARENT.R
\name{par_display_capabilities}
\alias{par_display_capabilities}
\title{Parse PARAMETERS packet for display capabilities}
\usage{
par_display_capabilities(value, key)
}
\arguments{
\item{value}{portion of payload to parse}

\item{key}{PARAMETERS key entry to use for guidance}
}
\description{
Parse PARAMETERS packet for display capabilities
}
\keyword{internal}
