% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_PARAMETERS_PARENT.R
\name{par_display_configuration}
\alias{par_display_configuration}
\title{Parse PARAMETERS packet for display configuration}
\usage{
par_display_configuration(value, key)
}
\arguments{
\item{value}{portion of payload to parse}

\item{key}{PARAMETERS key entry to use for guidance}
}
\description{
Parse PARAMETERS packet for display configuration
}
\keyword{internal}
