% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMETERS_payload_parsers_PARENT.R
\name{par_feature_capabilities}
\alias{par_feature_capabilities}
\title{Parse PARAMETERS packet for feature capabilities}
\usage{
par_feature_capabilities(value, key)
}
\arguments{
\item{value}{portion of payload to parse}

\item{key}{PARAMETERS key entry to use for guidance}
}
\description{
Parse PARAMETERS packet for feature capabilities
}
\keyword{internal}
