% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_PARAMETERS_PARENT.R
\name{par_feature_enable}
\alias{par_feature_enable}
\title{Parse PARAMETERS packet for enabled features}
\usage{
par_feature_enable(value, key)
}
\arguments{
\item{value}{portion of payload to parse}

\item{key}{PARAMETERS key entry to use for guidance}
}
\description{
Parse PARAMETERS packet for enabled features
}
\keyword{internal}
