% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_PARAMETERS_CHILD.R,
%   R/payload_PARAMETERS_PARENT.R
\name{par_wireless_firmware_version}
\alias{par_wireless_firmware_version}
\alias{par_firmware_version}
\title{Parse PARAMETERS packet for firmware version}
\usage{
par_wireless_firmware_version(value, key)

par_firmware_version(value, key)
}
\arguments{
\item{value}{portion of payload to parse}

\item{key}{PARAMETERS key entry to use for guidance}
}
\description{
Parse PARAMETERS packet for firmware version
}
\keyword{internal}
