% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_PARAMETERS_PARENT.R
\name{par_memory_size}
\alias{par_memory_size}
\title{Parse PARAMETERS packet for memory size}
\usage{
par_memory_size(value, key)
}
\arguments{
\item{value}{portion of payload to parse}

\item{key}{PARAMETERS key entry to use for guidance}
}
\description{
Parse PARAMETERS packet for memory size
}
\keyword{internal}
