% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_PARAMETERS_PARENT.R
\name{par_wireless_mode}
\alias{par_wireless_mode}
\title{Parse PARAMETERS packet for wireless mode}
\usage{
par_wireless_mode(value, key)
}
\arguments{
\item{value}{portion of payload to parse}

\item{key}{PARAMETERS key entry to use for guidance}
}
\description{
Parse PARAMETERS packet for wireless mode
}
\keyword{internal}
