% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMETERS_payload_parsers_CHILD.R,
%   R/PARAMETERS_payload_parsers_PARENT.R
\name{par_hrm_serial_number_h}
\alias{par_hrm_serial_number_h}
\alias{par_hrm_serial_number_l}
\alias{par_proximity_interval}
\alias{par_wireless_serial_number}
\title{Parse PARAMETERS packet for wireless serial number}
\usage{
par_hrm_serial_number_h(value, key)

par_hrm_serial_number_l(value, key)

par_proximity_interval(value, key)

par_wireless_serial_number(value, key)
}
\arguments{
\item{value}{portion of payload to parse}

\item{key}{PARAMETERS key entry to use for guidance}
}
\description{
Parse PARAMETERS packet for wireless serial number
}
\keyword{internal}
