% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{parse_IMU_C}
\alias{parse_IMU_C}
\title{Parse all IMU packets in a file}
\usage{
parse_IMU_C(imu_records, log, info, id, samp_rate, verbose)
}
\arguments{
\item{imu_records}{DataFrame with information about each packet}

\item{log}{RawVector containing all payload bytes}

\item{info}{The \code{sensorColumns} information of a \code{SENSOR_SCHEMA}
object}

\item{id}{integer. The \code{id} information of a \code{SENSOR_SCHEMA}
object}

\item{samp_rate}{integer. The IMU sampling rate.}

\item{verbose}{logical. Print updates to console?}
}
\description{
Parse all IMU packets in a file
}
\keyword{internal}
