% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_parse_info_txt.R
\name{parse_info_txt}
\alias{parse_info_txt}
\title{Parse the info component of a gt3x file}
\usage{
parse_info_txt(info, tz = "UTC", verbose, ...)
}
\arguments{
\item{info}{connection to the info.txt file}

\item{tz}{character. The timezone}

\item{verbose}{logical. Print updates to console?}

\item{...}{further arguments/methods. Currently unused.}
}
\description{
Parse the info component of a gt3x file
}
\keyword{internal}
