% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_log_bin.R
\name{parse_log_bin}
\alias{parse_log_bin}
\title{Parse the log component of a gt3x file}
\usage{
parse_log_bin(log_file, n_records, info, tz, verbose = FALSE,
  give_timestamp = TRUE, include = c("METADATA", "PARAMETERS",
  "SENSOR_SCHEMA", "BATTERY", "EVENT", "TAG", "ACTIVITY", "HEART_RATE_BPM",
  "HEART_RATE_ANT", "HEART_RATE_BLE", "LUX", "CAPSENSE", "EPOCH", "EPOCH2",
  "EPOCH3", "EPOCH4", "ACTIVITY2", "SENSOR_DATA"))
}
\arguments{
\item{log_file}{character. Path to the log.bin file}

\item{n_records}{integer. Length of the log}

\item{info}{result of \code{\link{parse_info_txt}}}

\item{tz}{character. The timezone to use}

\item{verbose}{logical. Print updates to console?}

\item{give_timestamp}{logical. Include timestamp in output?}

\item{include}{character. The PACKET types to parse}
}
\description{
Parse the log component of a gt3x file
}
\keyword{internal}
