% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_primary_accelerometer.R
\name{parse_primary_accelerometer}
\alias{parse_primary_accelerometer}
\title{Parse primary accelerometer packets}
\usage{
parse_primary_accelerometer(primary_records, log, info, tz, verbose = FALSE)
}
\arguments{
\item{primary_records}{data frame with information about the primary
accelerometer packets}

\item{log}{the gt3x raw data}

\item{info}{the result of \code{\link{parse_info_txt}}}

\item{tz}{character. The timezone to use}

\item{verbose}{logical. Print updates to console?}
}
\description{
Parse primary accelerometer packets
}
