% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_3x_payload_functions.R
\name{payload_parse}
\alias{payload_parse}
\title{Dispatch-like function for parsing different types of payloads}
\usage{
payload_parse(type, payload, info, tz = "UTC", parameters = NULL,
  schema = NULL, record_header = NULL, scale_factor,
  is_last_packet = FALSE)
}
\arguments{
\item{type}{The payload type}

\item{payload}{raw. The payload}

\item{info}{result of \code{\link{parse_info_txt}}}

\item{tz}{character. The timezone to use}

\item{parameters}{result of parsing the PARAMETERS packet}

\item{schema}{result of parsing the SENSOR_SCHEMA packet}

\item{record_header}{information for
\code{\link{payload_parse_sensor_data_25}} about the packet indices etc.}

\item{scale_factor}{scale_factor for ACTIVITY2 packets}

\item{is_last_packet}{logical. Is the current packet the last in the file?}
}
\description{
Dispatch-like function for parsing different types of payloads
}
\keyword{internal}
