% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finished_Payload_Functions.R
\name{payload_parse_activity2_26}
\alias{payload_parse_activity2_26}
\title{Parse ACTIVITY2 packet}
\usage{
payload_parse_activity2_26(payload, info, scale_factor,
  is_last_packet = FALSE)
}
\arguments{
\item{payload}{raw. The payload}

\item{info}{result of \code{\link{parse_info_txt}}}

\item{scale_factor}{scale_factor for ACTIVITY2 packets}

\item{is_last_packet}{logical. Is the current packet the last in the file?}
}
\description{
Parse ACTIVITY2 packet
}
\keyword{internal}
