% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unfinished_Payload_Functions.R
\name{payload_parse_activity_0}
\alias{payload_parse_activity_0}
\title{Parse ACTIVITY packet}
\usage{
payload_parse_activity_0(payload, info)
}
\arguments{
\item{payload}{raw. The payload}

\item{info}{result of \code{\link{parse_info_txt}}}
}
\description{
Parse ACTIVITY packet
}
\keyword{internal}
