% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finished_Payload_Functions.R
\name{payload_parse_battery_2}
\alias{payload_parse_battery_2}
\title{Parse BATTERY packet}
\usage{
payload_parse_battery_2(payload, info)
}
\arguments{
\item{payload}{raw. The payload}

\item{info}{result of \code{\link{parse_info_txt}}}
}
\description{
Parse BATTERY packet
}
\keyword{internal}
