% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Finished_Payload_Functions.R
\name{payload_parse_capsense_13}
\alias{payload_parse_capsense_13}
\title{Parse CAPSENSE packet}
\usage{
payload_parse_capsense_13(payload, info)
}
\arguments{
\item{payload}{raw. The payload}

\item{info}{result of \code{\link{parse_info_txt}}}
}
\description{
Parse CAPSENSE packet
}
\keyword{internal}
