% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unfinished_Payload_Functions.R
\name{payload_parse_epoch_15}
\alias{payload_parse_epoch_15}
\title{Parse EPOCH3 packet}
\usage{
payload_parse_epoch_15(payload, info)
}
\arguments{
\item{payload}{raw. The payload}

\item{info}{result of \code{\link{parse_info_txt}}}
}
\description{
Parse EPOCH3 packet
}
\keyword{internal}
