% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unfinished_Payload_Functions.R
\name{payload_parse_heart_rate_4}
\alias{payload_parse_heart_rate_4}
\title{Parse HEART_RATE_BPM packet}
\usage{
payload_parse_heart_rate_4(payload, info)
}
\arguments{
\item{payload}{raw. The payload}

\item{info}{result of \code{\link{parse_info_txt}}}
}
\description{
Parse HEART_RATE_BPM packet
}
\keyword{internal}
