% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unfinished_Payload_Functions.R
\name{payload_parse_lux_5}
\alias{payload_parse_lux_5}
\title{Parse LUX packet}
\usage{
payload_parse_lux_5(payload, info)
}
\arguments{
\item{payload}{raw. The payload}

\item{info}{result of \code{\link{parse_info_txt}}}
}
\description{
Parse LUX packet
}
\keyword{internal}
