% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMETERS_payload.R
\name{payload_parse_parameters_21}
\alias{payload_parse_parameters_21}
\title{Master function to parse PARAMETERS packets}
\usage{
payload_parse_parameters_21(payload, info, tz = "UTC")
}
\arguments{
\item{payload}{raw. The payload}

\item{info}{result of \code{\link{parse_info_txt}}}

\item{tz}{character. The timezone to use.}
}
\description{
Master function to parse PARAMETERS packets
}
\keyword{internal}
