% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_SENSOR_DATA.R
\name{payload_parse_sensor_data_25}
\alias{payload_parse_sensor_data_25}
\title{Parse SENSOR_DATA packet}
\usage{
payload_parse_sensor_data_25(payload, parameters, schema)
}
\arguments{
\item{payload}{raw. The packet payload}

\item{parameters}{a PARAMETERS object}

\item{schema}{a SENSOR_SCHEMA object}
}
\description{
This is primarily a wrapper for \code{\link{legacy_payload_parse_sensor_data_25C}}
}
\keyword{internal}
