% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{payload_parse_sensor_data_25C}
\alias{payload_parse_sensor_data_25C}
\title{Parse SENSOR_DATA packet in c++}
\usage{
payload_parse_sensor_data_25C(payload, schema)
}
\arguments{
\item{payload}{raw. The payload}

\item{schema}{result of parsing the SENSOR_SCHEMA packet}
}
\description{
Parse SENSOR_DATA packet in c++
}
\keyword{internal}
