% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SENSOR_SCHEMA_payload.R
\name{payload_parse_sensor_schema_24}
\alias{payload_parse_sensor_schema_24}
\title{Parse metadata from a SENSOR_SCHEMA packet}
\usage{
payload_parse_sensor_schema_24(payload, info)
}
\arguments{
\item{payload}{raw. The payload}

\item{info}{result of \code{\link{parse_info_txt}}}
}
\description{
Parse metadata from a SENSOR_SCHEMA packet
}
\keyword{internal}
