% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unfinished_Payload_Functions.R
\name{payload_parse_tag_7}
\alias{payload_parse_tag_7}
\title{Parse TAG packet}
\usage{
payload_parse_tag_7(payload, info)
}
\arguments{
\item{payload}{raw. The payload}

\item{info}{result of \code{\link{parse_info_txt}}}
}
\description{
Parse TAG packet
}
\keyword{internal}
