% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_record_set.R
\name{post_process}
\alias{post_process}
\title{Apply post-processing to the result of parsing packets of a certain type}
\usage{
post_process(result)
}
\arguments{
\item{result}{the result to process}
}
\description{
Apply post-processing to the result of parsing packets of a certain type
}
\keyword{internal}
