% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_progC}
\alias{print_progC}
\title{Print progress updates while parsing packets in C++}
\usage{
print_progC(n, label)
}
\arguments{
\item{n}{percentage progress}

\item{label}{the packet type, as character}
}
\description{
Print progress updates while parsing packets in C++
}
\keyword{internal}
