% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMETERS_payload.R
\name{process_parameters}
\alias{process_parameters}
\title{Second-level function for parsing PARAMETERS packets}
\usage{
process_parameters(payload_chunk, tz)
}
\arguments{
\item{tz}{character. The timezone to use}
}
\description{
Dispatches to a third-level function (e.g. \code{\link{par_battery_state}})
depending on the packet address and identifier
}
\keyword{internal}
