% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_record_set.R
\name{process_record_set}
\alias{process_record_set}
\title{Parse the packets of a certain type}
\usage{
process_record_set(record_set, log, tz, info, give_timestamp,
  parameters = NULL, schema = NULL, verbose = FALSE,
  do_post_process = TRUE)
}
\arguments{
\item{record_set}{data frame. The record entries for the packet type.}

\item{log}{raw. The data from log.bin}

\item{tz}{character. The timezone to use}

\item{info}{result of \code{\link{parse_info_txt}}}

\item{give_timestamp}{logical. Include timestamp in output?}

\item{parameters}{result of parsing the PARAMETERS packet}

\item{schema}{result of parsing the SENSOR_SCHEMA packet}

\item{verbose}{logical. Print updates to console?}

\item{do_post_process}{logical. Divert to \code{\link{post_process}} before
returning?}
}
\description{
Parse the packets of a certain type
}
\keyword{internal}
