% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGread_Data.R
\docType{data}
\name{raw_to_collapse}
\alias{raw_to_collapse}
\title{Primary accelerometer data to collapse}
\format{A data frame with 24000 rows and 3 variables:
\describe{
  \item{Accelerometer X}{Primary accelerometer x-axis data, in G}
  \item{Accelerometer Y}{Primary accelerometer y-axis data, in G}
  \item{Accelerometer Z}{Primary accelerometer z-axis data, in G}
}}
\usage{
raw_to_collapse
}
\description{
A partially-processed primary accelerometer dataset ready to be collapsed
from raw samples to one-second summaries.
}
\keyword{datasets}
