% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AG_raw_block.R
\name{read_AG_raw_block}
\alias{read_AG_raw_block}
\title{Read a large raw file in blocks}
\usage{
read_AG_raw_block(file, output_window_secs, calibrate, verbose, skip, meta,
  timer, ...)
}
\arguments{
\item{file}{A character scalar giving path to primary accelerometer file}

\item{output_window_secs}{the desired epoch length; defaults to one second}

\item{calibrate}{logical. Perform autocalibration using \link[GGIR]{g.calibrate}}

\item{verbose}{A logical scalar: Print processing updates?}

\item{skip}{Header length: Number of rows to skip when reading the file}

\item{meta}{information from \code{\link{get_raw_file_meta}}}

\item{timer}{initial time passed from initial call to \code{\link{read_AG_raw}}}

\item{...}{Arguments passed to \code{\link{read.csv}} in
\code{\link{check_columns}}}
}
\description{
Read a large raw file in blocks
}
\keyword{internal}
