% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_info.R
\name{info_type1}
\alias{info_type1}
\alias{read_gt3x_info}
\title{Retrieve information from the gt3x zip archive}
\usage{
info_type1(file, tz, verbose)

read_gt3x_info(file, tz, verbose)
}
\arguments{
\item{file}{character. Path to the file}

\item{tz}{character. The timezone to use}

\item{verbose}{logical. Print updates to console?}
}
\description{
Retrieve information from the gt3x zip archive
}
\keyword{internal}
