% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_record.R
\name{read_record}
\alias{read_record}
\title{High-level function to read a packet of data from a gt3x file}
\usage{
read_record(record_header, log, tz = "UTC", info, give_timestamp = TRUE,
  parameters = NULL, schema = NULL, scale_factor, ...)
}
\arguments{
\item{record_header}{data frame containing information about the packet
indices etc.}

\item{log}{raw. The data from log.bin}

\item{tz}{character. The timezone to use}

\item{info}{result of \code{\link{parse_info_txt}}}

\item{give_timestamp}{logical. Include timestamp in output?}

\item{parameters}{result of parsing the PARAMETERS packet}

\item{schema}{result of parsing the SENSOR_SCHEMA packet}

\item{scale_factor}{scale_factor for ACTIVITY2 packets}

\item{...}{further arguments passed to \code{\link{payload_parse}}}
}
\description{
High-level function to read a packet of data from a gt3x file
}
\keyword{internal}
