% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reintegrate_Utils.R
\name{reint_wrap}
\alias{reint_wrap}
\title{Run a reintegration operation on a data frame with a \code{block} indicator}
\usage{
reint_wrap(ag, input_vars, fun)
}
\arguments{
\item{ag}{data frame with a \code{block} column to indicate groupings for the
reintegration operation}

\item{input_vars}{character vector of column names on which to reintegrate}

\item{fun}{the function(s) to perform}
}
\description{
Run a reintegration operation on a data frame with a \code{block} indicator
}
\keyword{internal}
