% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reintegrate2.R
\name{reintegrate2}
\alias{reintegrate2}
\title{A fresh approach to reintegrating ActiGraph data}
\usage{
reintegrate2(ag, target_sec, time_var = "Timestamp")
}
\arguments{
\item{ag}{a data frame to reintegrate}

\item{target_sec}{the desired epoch length of the output. Starting epoch
length will be determined automatically}

\item{time_var}{The name of the column containing POSIX-formatted timestamp
information}
}
\description{
This is a \code{tidy} way of reintegrating, which is both more
straightforward and more concise. Runtime implications are uncertain, and
some functionality (notably the \code{direction} argument) has been
abandoned. Only forward reintegration is supported here (consistent with
ActiLife precedent).
}
\examples{
test_file <- system.file(
  "extdata", "example1sec.csv", package = "AGread"
)

ag <- read_AG_counts(test_file, skip = 11)

# Old Method:

  old_result <- reintegrate(
    ag = ag,
    to = 60,
    time_var = "Timestamp",
    direction = c("forwards")
  )

# New Method:

  new_result <- reintegrate2(ag, 60)
}
