% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reintegrate_Helper_Functions.R
\name{rm_trail_na}
\alias{rm_trail_na}
\title{Remove trailing rows with missing values from data frame}
\usage{
rm_trail_na(ag)
}
\arguments{
\item{ag}{A data frame on which to perform the operation}
}
\description{
Remove trailing rows with missing values from data frame
}
\keyword{internal}
