% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SENSOR_DATA_payload.R
\name{row_test}
\alias{row_test}
\title{Test whether a SENSOR_DATA packet has the expected number of rows}
\usage{
row_test(all_indices, payload, expected_rows)
}
\arguments{
\item{all_indices}{matrix. Indices of expected payload entries}

\item{payload}{raw. The payload}

\item{expected_rows}{integer. The expected number of rows}
}
\description{
Test whether a SENSOR_DATA packet has the expected number of rows
}
\keyword{internal}
