% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SENSOR_SCHEMA_payload.R
\name{schema_meta}
\alias{schema_meta}
\title{Parse metadata from a SENSOR_SCHEMA packet}
\usage{
schema_meta(payload)
}
\arguments{
\item{payload}{raw. The payload}
}
\description{
Parse metadata from a SENSOR_SCHEMA packet
}
\keyword{internal}
