% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_Interpolate2.R
\name{sensor_resample}
\alias{sensor_resample}
\alias{res_interpolate}
\title{Resample a sensor data stream}
\usage{
sensor_resample(original_samples, target_frequency, method = "interpolate",
  ...)

res_interpolate(original_samples, target_frequency, ...)
}
\arguments{
\item{original_samples}{numeric. the original data stream}

\item{target_frequency}{numeric. the target sampling frequency}

\item{method}{character. Resampling method to apply. Currently accepts only
\code{interpolate}.}

\item{...}{additional arguments to \code{method}}
}
\value{
An appropriately up/down-sampled data stream (numeric vector)
}
\description{
This function was originally developed to correct sensor packets when sampled
more or fewer times than expected over the time interval. However, this
function may have more far-reaching applications.
}
\examples{
set.seed(14)
target_frequency <- 100

## Downsample
   original_samples <- sample(
     seq(1.3,2.4,0.12), 101, replace = TRUE
   )
   sensor_resample(original_samples, target_frequency)

## Upsample
   original_samples <- original_samples[1:99]
   sensor_resample(original_samples, target_frequency)
}
\keyword{internal}
