% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{seq_interpolate}
\alias{seq_interpolate}
\title{One-dimensional linear interpolation}
\usage{
seq_interpolate(from, to, ...)
}
\arguments{
\item{from}{starting value}

\item{to}{ending value}

\item{...}{further arguments passed to \code{seq}}
}
\description{
One-dimensional linear interpolation
}
\keyword{internal}
