% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_parse_log_bin_utils.R
\name{setup_payload}
\alias{setup_payload}
\title{Retrieve payload data for a single packet}
\usage{
setup_payload(record_header, log)
}
\arguments{
\item{record_header}{data frame containing information about the packet
indices etc.}

\item{log}{raw. The data from log.bin}
}
\description{
This function includes a call to \code{\link{checksumC}} and will break if
the result is not as expected
}
\keyword{internal}
