% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_log_bin_utils.R
\name{sort_records}
\alias{sort_records}
\title{Sort chronologically-ordered record headers by type, for parsing one type at
a time}
\usage{
sort_records(record_headers)
}
\arguments{
\item{record_headers}{the record headers to sort}
}
\description{
Sort chronologically-ordered record headers by type, for parsing one type at
a time
}
\keyword{internal}
