% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_header.R
\name{special_header}
\alias{special_header}
\title{Individually process a specific record_header}
\usage{
special_header(record_headers, type, log, tz, info, give_timestamp,
  verbose = FALSE, do_post_process = FALSE)
}
\arguments{
\item{record_headers}{list of record header information for the various
packet types}

\item{type}{character. The type code for the target record header}

\item{log}{raw. The data from log.bin}

\item{tz}{character. The timezone to use}

\item{info}{result of \code{\link{parse_info_txt}}}

\item{give_timestamp}{logical. Include timestamp in output?}

\item{verbose}{logical. Print updates to console?}

\item{do_post_process}{logical. Divert to \code{\link{post_process}} before
returning?}
}
\description{
Individually process a specific record_header
}
\keyword{internal}
