% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_record_set.R
\name{test_sensor_records}
\alias{test_sensor_records}
\title{Check to see if SENSOR_DATA packets have the expected number of records}
\usage{
test_sensor_records(record_set, schema)
}
\arguments{
\item{record_set}{data frame. Information about the SENSOR_DATA packets (one
row per packet)}

\item{schema}{result of parsing the SENSOR_SCHEMA packet}
}
\description{
Check to see if SENSOR_DATA packets have the expected number of records
}
\keyword{internal}
