% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Function.R
\name{test_times}
\alias{test_times}
\title{Test whether first columns contain date/time information}
\usage{
test_times(AG, verbose = FALSE)
}
\arguments{
\item{AG}{A data frame of partially-processed data}

\item{verbose}{A logical scalar: Print processing updates?}
}
\description{
Test whether first columns contain date/time information
}
\keyword{internal}
