% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Counts.R
\name{test_VM}
\alias{test_VM}
\title{Test whether a vector magnitude column is present}
\usage{
test_VM(AG, verbose = FALSE)
}
\arguments{
\item{AG}{A data frame of partially-processed data}

\item{verbose}{A logical scalar: Print processing updates?}
}
\description{
Test whether a vector magnitude column is present
}
\keyword{internal}

