% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp_recalc.R
\name{timestamp_recalc}
\alias{timestamp_recalc}
\title{Re-calculate IMU timestamps after parsing and interpolating SENSOR_DATA
packets}
\usage{
timestamp_recalc(timestamps, tz, verbose, samp_rate, label)
}
\arguments{
\item{timestamps}{POSIX-formatted vector of timestamps}

\item{tz}{character. The timezone to use}

\item{verbose}{logical. Print updates to console?}

\item{samp_rate}{integer. The sampling rate.}

\item{label}{character. The packet type}
}
\description{
Re-calculate IMU timestamps after parsing and interpolating SENSOR_DATA
packets
}
\keyword{internal}
