% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_utils.R
\name{update_key}
\alias{update_key}
\title{Final formatting for a PARAMETERS result}
\usage{
update_key(key, value)
}
\arguments{
\item{key}{the key entry to use as a shell}

\item{value}{the value to append}
}
\description{
Final formatting for a PARAMETERS result
}
\keyword{internal}
