% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_log_bin_utils.R
\name{validate_include}
\alias{validate_include}
\title{Validate the values provided for the \code{input} argument in
\code{\link{read_gt3x}}}
\usage{
validate_include(include, verbose = FALSE, choices = gt3x_packets())
}
\arguments{
\item{include}{character. The PACKET types to parse}

\item{verbose}{logical. Print updates to console?}

\item{choices}{character. The packet types to choose from.}
}
\description{
Validate the values provided for the \code{input} argument in
\code{\link{read_gt3x}}
}
\keyword{internal}
