% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{zero2one}
\alias{zero2one}
\title{Interpolation-specific sequencer}
\usage{
zero2one(samples)
}
\arguments{
\item{samples}{NumericVector. A data stream.}
}
\description{
Interpolation-specific sequencer
}
\keyword{internal}
