% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_fill.R
\name{zero_fill}
\alias{zero_fill}
\alias{zero_fill.list}
\alias{zero_fill.data.frame}
\title{Fill in missing data packets}
\usage{
zero_fill(records, index, ...)

\method{zero_fill}{list}(records, index, info = NULL, schema = NULL, ...)

\method{zero_fill}{data.frame}(records, index, info = NULL, schema = NULL,
  verbose = FALSE, ...)
}
\arguments{
\item{records}{list of parsed packets}

\item{index}{index of record after which packets are missing}

\item{...}{further arguments}

\item{info}{result of \code{\link{parse_info_txt}}}

\item{schema}{result of parsing the SENSOR_SCHEMA packet}
}
\description{
Fill in missing data packets
}
\keyword{internal}
