% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_pad.R
\name{zero_pad}
\alias{zero_pad}
\alias{zero_pad.list}
\alias{zero_pad.RAW}
\title{Fill in missing data packets at start/end of file}
\usage{
zero_pad(object, ...)

\method{zero_pad}{list}(object, index, info = NULL, schema = NULL, ...)

\method{zero_pad}{RAW}(object, info, verbose, ...)
}
\arguments{
\item{object}{the packet stream}

\item{...}{further arguments passed to methods and \code{\link{zero_pad}}}

\item{info}{the content of \code{info.txt}}

\item{verbose}{logical. Print updates to console?}
}
\description{
This function checks whether a packet stream's starting and ending timestamps
correspond with what's expected from \code{info.txt}.
}
\keyword{internal}
