% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_Reading_Helper_Functions.R
\name{AG_collapse}
\alias{AG_collapse}
\title{Collapse primary accelerometer data}
\usage{
AG_collapse(AG, output_window_secs = 1, samp_freq)
}
\arguments{
\item{AG}{a dataframe of raw primary accelerometer data}

\item{output_window_secs}{the desired epoch length; defaults to one second}

\item{samp_freq}{The sampling frequency}
}
\description{
Collapse primary accelerometer data
}
\examples{
data(raw_to_collapse)
AG_collapse(raw_to_collapse, 1, 80)

}
