% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Function.R
\name{AG_meta}
\alias{AG_meta}
\title{Extract meta-data from file header}
\usage{
AG_meta(file, verbose = FALSE, ...)
}
\arguments{
\item{file}{A character scalar giving path to an automatically-generated csv
file with count values}

\item{verbose}{A logical scalar: Print processing updates?}

\item{...}{Further arguments passed to \code{read.csv} and \code{fread}}
}
\description{
Extract meta-data from file header
}
\keyword{internal}
