% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{agd_format}
\alias{agd_format}
\alias{agd_date_string}
\alias{agd_vector_magnitude}
\title{Format ActiGraph agd data after initial reading}
\usage{
agd_format(AG, return, tz, file)

agd_date_string(timestamp, tz)

agd_vector_magnitude(AG, expected = paste0("Axis", 1:3))
}
\arguments{
\item{AG}{list. The result of initial reading}

\item{return}{character. The desired elements to return (see details)}

\item{tz}{character. The desired timezone}

\item{file}{character. Path to the agd file}

\item{expected}{character. The expected names of the triaxial variables}
}
\description{
Format ActiGraph agd data after initial reading
}
\keyword{internal}
