% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGread-deprecated.R, R/Read_Counts.R
\name{AGread-deprecated}
\alias{AGread-deprecated}
\alias{get_duration}
\title{Deprecated functions in package \pkg{AGread}.}
\usage{
get_duration(timer)
}
\description{
These are functions that are deprecated because they have been
  moved elsewhere, primarily to the PAutilities package. Use
  \code{help("-deprecated")} to access documentation for deprecated
  functions.
}
\section{\code{get_duration}}{

For \code{get_duration}, use \code{\link[PAutilities]{get_duration}}
}

\keyword{internal}
