% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_Reading_Helper_Functions.R
\name{check_columns}
\alias{check_columns}
\title{Check if the primary accelerometer file is formatted correctly}
\usage{
check_columns(file)
}
\arguments{
\item{file}{A character scalar giving path to primary accelerometer file}
}
\description{
\code{check_columns} returns a logical scalar indicating whether there is a
 formatting issue with the file passed as the argument. A value of TRUE
 indicates the test has passed, whereas FALSE indicates an issue.
}
\examples{
\dontrun{
raw_file <-
    system.file("extdata",
    "TestID_LeftWrist_RAW.csv",
    package = "AGread")

check_columns(raw_file)
}

}
