% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gaps.R, R/check_gaps_utils.R
\name{check_gaps}
\alias{check_gaps}
\alias{check_gaps.RAW}
\alias{get_latch_index}
\alias{sleep_latch}
\alias{trailing_zeroes}
\alias{get_missing_times}
\title{Examine a packet stream for gaps}
\usage{
check_gaps(object, ...)

\method{check_gaps}{RAW}(object, set, info, events, ...)

get_latch_index(missing_times, reference_times, tz)

sleep_latch(object, tz, info, events)

trailing_zeroes(object, tz, info)

get_missing_times(object, tz, info)
}
\arguments{
\item{object}{the packet stream}

\item{...}{further arguments passed to methods}

\item{set}{the set of record headers corresponding to each packet}

\item{info}{the result of \code{\link{parse_info_txt}}}

\item{events}{the result of parsing EVENTS packets}

\item{missing_times}{vector of missing timestamps for which to identify a
latch index}

\item{reference_times}{vector of reference timestamps for use in determining
the latch index}

\item{tz}{the timezone}
}
\description{
This function also checks the beginning and end of the stream to see if any
padding is necessary based on \code{info.txt}
}
\keyword{internal}
