% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/check_gaps.R,
%   R/check_gaps_utils.R
\name{latch_accel}
\alias{latch_accel}
\alias{get_latch_index}
\alias{get_latch_values}
\alias{get_latch_entries}
\alias{check_gaps}
\alias{check_gaps.RAW}
\alias{get_missing_times}
\alias{sleep_latch}
\title{Examine a packet stream for gaps}
\usage{
latch_accel(vector_size, accel_input, samp_rate)

get_latch_index(missing_times, reference_times)

get_latch_values(sleeps, RAW)

get_latch_entries(samp_rate, timestamps, accel_x, accel_y, accel_z,
  return_empty = FALSE)

check_gaps(object, ...)

\method{check_gaps}{RAW}(object, set, info, events, ...)

get_missing_times(object, tz, info)

sleep_latch(object, info, events, tz)
}
\arguments{
\item{vector_size}{int. The size of the final vector}

\item{accel_input}{NumericVector. The acceleromter values to reference for
latching}

\item{missing_times}{vector of missing timestamps for which to identify a
latch index}

\item{reference_times}{vector of reference timestamps for use in determining
the latch index}

\item{sleeps}{DataFrame containing idle sleep mode information}

\item{RAW}{DataFrame containing raw acceleration data}

\item{object}{the packet stream}

\item{...}{further arguments passed to methods}

\item{set}{the set of record headers corresponding to each packet}

\item{info}{the result of \code{\link{parse_info_txt}}}

\item{events}{the result of parsing EVENTS packets}

\item{tz}{the timezone}
}
\description{
This function also checks the beginning and end of the stream to see if any
padding is necessary based on \code{info.txt}
}
\keyword{internal}
