% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gaps.R
\name{check_gaps}
\alias{check_gaps}
\alias{check_gaps.RAW}
\title{Examine a packet stream for gaps}
\usage{
check_gaps(object, ...)

\method{check_gaps}{RAW}(object, info, ...)
}
\arguments{
\item{object}{the packet stream}

\item{...}{further arguments passed to methods and \code{\link{zero_pad}}}

\item{info}{the content of \code{info.txt}}
}
\description{
This function also checks the beginning and end of the stream to see if any
padding is necessary based on \code{info.txt}
}
\keyword{internal}
