% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_id}
\alias{check_id}
\title{Check sensor payload ID prior to parsing the packet}
\usage{
check_id(x, id)
}
\arguments{
\item{x}{the raw payload}

\item{id}{integer. The \code{SENSOR_SCHEMA} id}
}
\description{
Check sensor payload ID prior to parsing the packet
}
\keyword{internal}
