% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Counts_Utils.R
\name{check_inc}
\alias{check_inc}
\title{Test whether inclinometer variables have been correctly assigned}
\usage{
check_inc(AG, verbose = FALSE)
}
\arguments{
\item{AG}{the monitor data to check}
}
\description{
Test whether inclinometer variables have been correctly assigned
}
\keyword{internal}
