% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_record_set.R
\name{collapse_records}
\alias{collapse_records}
\alias{generic_record_collapse}
\title{Dispatch-like function for combining packet records of a certain type}
\usage{
collapse_records(records, label, verbose = FALSE)

generic_record_collapse(records, label)
}
\arguments{
\item{records}{The records to combine}

\item{label}{The record type}

\item{verbose}{logical. Print updates to console?}
}
\description{
Dispatch-like function for combining packet records of a certain type
}
\keyword{internal}
