% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dev_parse_activity}
\alias{dev_parse_activity}
\title{Parse a set of ACTIVITY packets}
\usage{
dev_parse_activity(packets, packet_no, zero_packet, samp_rate, scale_factor)
}
\arguments{
\item{packets}{list of packets}

\item{packet_no}{IntegerVector indicating which index of \code{packets} to
use for each second of expected output. Values of -1 indicate a latch to
the previous index}

\item{zero_packet}{list containing a properly-formatted packet pre-filled
with values of zero (used for USB connection events)}

\item{latch_packets}{list of empty packets to be filled during latch periods}
}
\description{
Parse a set of ACTIVITY packets
}
\keyword{internal}
