% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_idle.R
\name{flag_idle}
\alias{flag_idle}
\title{Flag Idle Sleep Times}
\usage{
flag_idle(RAW, EVENT, verbose = FALSE)
}
\arguments{
\item{RAW}{packet from \code{\link{parse_log_bin}}}

\item{EVENT}{packet from \code{\link{parse_log_bin}}}

\item{verbose}{logical. Print updates to console?}
}
\description{
Flag Idle Sleep Times
}
\keyword{internal}
