% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reintegrate_Utils.R
\name{get_blocks}
\alias{get_blocks}
\title{Assign blocks to data stream for reintegration}
\usage{
get_blocks(ag, time_var, to, start_epoch, block_size, direction)
}
\arguments{
\item{ag}{A data frame to reintegrate}

\item{time_var}{The name of the column containing POSIX-formatted timestamp
information}

\item{to}{The epoch length desired. Starting epoch length will be determined
automatically.}

\item{start_epoch}{The initial epoch length of the data being reintegrated}

\item{block_size}{The number of rows of data included in each reintegrated
epoch}

\item{direction}{The direction of reintegration, i.e. whether a timestamp
refers to the timespan after the previous data point ("backwards"), or
before the next data point ("forwards").}
}
\description{
Assign blocks to data stream for reintegration
}
\keyword{internal}
