% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_Counts.R
\name{get_duration}
\alias{get_duration}
\title{Provide the run time of processing}
\usage{
get_duration(t)
}
\arguments{
\item{t}{The initial time}
}
\description{
Provide the run time of processing
}
\keyword{internal}
