% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_headersC}
\alias{get_headersC}
\title{Collect informtion about the packets stored in log.bin}
\usage{
get_headersC(x)
}
\arguments{
\item{x}{RawVector. The contents of log.bin}
}
\description{
Collect informtion about the packets stored in log.bin
}
\keyword{internal}
