% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_3x_read.R
\name{get_primary_accel_scale}
\alias{get_primary_accel_scale}
\title{Retrieve the primary accelerometer scale factor}
\usage{
get_primary_accel_scale(info)
}
\value{
the scale factor, as integer
}
\description{
Retrieve the primary accelerometer scale factor
}
\keyword{internal}
