% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/File_Reading_Helper_Functions.R
\name{get_raw_file_meta}
\alias{get_raw_file_meta}
\title{Get file metadata (sampling frequency and timestamps) for primary accelerometer}
\usage{
get_raw_file_meta(file)
}
\arguments{
\item{file}{character scalar giving path to primary accelerometer file}
}
\description{
Get file metadata (sampling frequency and timestamps) for primary accelerometer
}
\examples{
\dontrun{
raw_file <-
    system.file("extdata",
    "TestID_LeftWrist_RAW.csv",
    package = "AGread")

get_raw_file_meta(raw_file)
}

}
