% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_times}
\alias{get_times}
\title{Determine the expected timestamps for primary accelerometer output}
\usage{
get_times(start, end, samp_rate, extra_packet = FALSE)
}
\arguments{
\item{start}{The file start time}

\item{end}{The file end time}

\item{samp_rate}{int. The sampling rate}

\item{extra_packet}{bool. Add an extra packet at the end?}
}
\description{
Determine the expected timestamps for primary accelerometer output
}
\keyword{internal}
