% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latch_gt3x.R
\name{latch_gt3x}
\alias{latch_gt3x}
\alias{get_sleep}
\title{Replace imputed zeroes by carrying forward last non-zero value}
\usage{
latch_gt3x(
  AG,
  x_var = "X",
  y_var = "Y",
  z_var = "Z",
  verbose = FALSE,
  flag_idle_sleep = FALSE,
  ...
)

get_sleep(AG, x_var = "X", y_var = "Y", z_var = "Z", ...)
}
\arguments{
\item{AG}{data frame of ActiGraph data (could be an \code{activity_df} object
from the \code{read.gt3x} package)}

\item{x_var, y_var, z_var}{character scalars naming the three columns of raw
acceleration data in \code{AG}}

\item{verbose}{logical. Print updates to console?}

\item{flag_idle_sleep}{should recorded idle sleep times be tagged?}

\item{...}{arguments passed to \code{get_sleep}}

\item{time_var}{character scalar naming the column of \code{AG} containing
POSIX-formatted timestamps}
}
\value{
An updated data frame with latched values rather than zeroes for
  missing data
}
\description{
Replace imputed zeroes by carrying forward last non-zero value
}
\examples{
\donttest{
file_3x <- system.file(
  "extdata", "example.gt3x", package = "AGread"
)
AG <- read_gt3x(file_3x)$RAW
head(latch_gt3x(AG, "Accelerometer_X", "Accelerometer_Y", "Accelerometer_Z"))
}

}
\keyword{internal}
