% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latch_value.R
\name{latch_value}
\alias{latch_value}
\title{Latch a data packet to the last payload sample of the previous packet}
\usage{
latch_value(records, index)
}
\arguments{
\item{records}{list of parsed packets}

\item{index}{index of record after which packets are missing}
}
\description{
Latch a data packet to the last payload sample of the previous packet
}
\keyword{internal}
