% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{message_update}
\alias{message_update}
\title{Print processing updates to console}
\usage{
message_update(message_number, is_message = FALSE, file, dur, vm_variables, n,
  filter_hz)
}
\arguments{
\item{message_number}{The message number to print}

\item{is_message}{A logical scalar: Print as message?}

\item{file}{A character scalar: name of file being processed}

\item{vm_variables}{variables being used to calculate vector magnitude}

\item{n}{Number of replications to be performed for magnetometer direction classification}

\item{filter_hz}{The cutoff for the low-pass filter}
}
\description{
Print processing updates to console
}
\keyword{internal}

