% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_bin_type1.R,
%   R/read_gt3x_parse_log_bin.R, R/read_gt3x_parse_log_bin_utils.R
\name{legacy_bin_type1}
\alias{legacy_bin_type1}
\alias{legacy_parse}
\alias{parse_log_bin}
\alias{validate_parser}
\title{Parse the log component of a gt3x file}
\usage{
legacy_bin_type1(log, tz, verbose = FALSE, include, info)

legacy_parse(file, ...)

parse_log_bin(log_file, info, tz = "UTC", verbose = FALSE,
  include = .packets, parser, file)

validate_parser(parser)
}
\arguments{
\item{tz}{character. The timezone to use}

\item{verbose}{logical. Print updates to console?}

\item{include}{character. The PACKET types to parse}

\item{info}{result of \code{\link{parse_info_txt}}}

\item{file}{internal list object containing information about the zip archive}

\item{...}{for internal use. Arguments passed to the specific parser from the
\code{parse_log_bin} shell}

\item{log_file}{character. Path to the log.bin file}

\item{parser}{the parsing scheme to use, either \code{legacy} or \code{dev}.
The former runs slower but includes more extensive checks to ensure
alignment with \code{RAW.csv} and \code{IMU.csv} files. The latter runs
faster and has also been checked for alignment with \code{RAW.csv} and
\code{IMU.csv} files, but not as strictly. For example, rounding is not
performed by \code{parser="dev"}.}
}
\description{
Parse the log component of a gt3x file
}
\keyword{internal}
