% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_log_bin.R
\name{parse_log_bin}
\alias{parse_log_bin}
\title{Parse the log component of a gt3x file}
\usage{
parse_log_bin(
  log_file,
  file_3x_len,
  info,
  tz = "UTC",
  verbose = FALSE,
  include = gt3x_packets()
)
}
\arguments{
\item{log_file}{character. Path to the log.bin file}

\item{file_3x_len}{length of log.bin, obtained from the \code{file_3x} object
from the parent function \code{\link{read_gt3x}}}

\item{info}{result of \code{\link{parse_info_txt}}}

\item{tz}{character. The timezone to use}

\item{verbose}{logical. Print updates to console?}

\item{include}{character. The PACKET types to parse}
}
\description{
Parse the log component of a gt3x file
}
\keyword{internal}
