% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_packet_set_ACTIVITY2.R,
%   R/parse_packet_set_CAPSENSE&METADATA.R,
%   R/parse_packet_set_SENSOR_DATA.R, R/parse_packet_set_SENSOR_SCHEMA.R,
%   R/parse_packet_set__generic_&_simple_methods.R
\name{parse_packet_set.ACTIVITY2}
\alias{parse_packet_set.ACTIVITY2}
\alias{parse_packet_set.CAPSENSE}
\alias{parse_packet_set.METADATA}
\alias{parse_packet_set.SENSOR_DATA}
\alias{parse_packet_set.SENSOR_SCHEMA}
\alias{parse_packet_set}
\alias{parse_packet_set.default}
\alias{parse_packet_set.PARAMETERS}
\alias{parse_packet_set.BATTERY}
\alias{parse_packet_set.EVENT}
\title{Parse all packets of a given type}
\usage{
\method{parse_packet_set}{ACTIVITY2}(set, log, tz = "UTC", verbose = FALSE,
  info, ...)

\method{parse_packet_set}{CAPSENSE}(set, log, tz = "UTC", verbose = FALSE,
  ...)

\method{parse_packet_set}{METADATA}(set, log, tz = "UTC", verbose = FALSE,
  ...)

\method{parse_packet_set}{SENSOR_DATA}(set, log, tz = "UTC",
  verbose = FALSE, schema, ...)

\method{parse_packet_set}{SENSOR_SCHEMA}(set, log, tz = "UTC",
  verbose = FALSE, ...)

parse_packet_set(set, log, tz = "UTC", verbose = FALSE, ...)

\method{parse_packet_set}{default}(set, log, tz = "UTC", verbose = FALSE,
  ...)

\method{parse_packet_set}{PARAMETERS}(set, log, tz = "UTC", verbose = FALSE,
  ...)

\method{parse_packet_set}{BATTERY}(set, log, tz = "UTC", verbose = FALSE,
  ...)

\method{parse_packet_set}{EVENT}(set, log, tz = "UTC", verbose = FALSE, ...)
}
\arguments{
\item{set}{the set of record headers corresponding to each packet}

\item{log}{the raw data from \code{log.bin}}

\item{tz}{character. The timezone to use}

\item{verbose}{logical. Print updates to console?}

\item{info}{the result of \code{\link{parse_info_txt}}}

\item{...}{further arguments passed to methods}
}
\description{
Parse all packets of a given type
}
\keyword{internal}
