% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SENSOR_DATA_payload.R
\name{payload_parse_sensor_data_25}
\alias{payload_parse_sensor_data_25}
\title{Parse SENSOR_DATA packet}
\usage{
payload_parse_sensor_data_25(payload, parameters, schema)
}
\arguments{
\item{payload}{raw. The payload}

\item{parameters}{result of parsing the PARAMETERS packet}

\item{schema}{result of parsing the SENSOR_SCHEMA packet}
}
\description{
This is primarily a wrapper for \code{\link{payload_parse_sensor_data_25C}}
}
\keyword{internal}
