% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SENSOR_DATA_payload.R
\name{payload_parse_sensor_data_25}
\alias{payload_parse_sensor_data_25}
\title{Parse SENSOR_DATA packet}
\usage{
payload_parse_sensor_data_25(payload, info, parameters, schema, record_header)
}
\arguments{
\item{payload}{raw. The payload}

\item{info}{result of \code{\link{parse_info_txt}}}

\item{parameters}{result of parsing the PARAMETERS packet}

\item{schema}{result of parsing the SENSOR_SCHEMA packet}

\item{record_header}{data frame with information about the record in question}
}
\description{
Parse SENSOR_DATA packet
}
\keyword{internal}
