% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{payload_parse_sensor_data_25C}
\alias{payload_parse_sensor_data_25C}
\title{Parse SENSOR_DATA packet in c++}
\usage{
payload_parse_sensor_data_25C(payload, info, id)
}
\arguments{
\item{payload}{raw vector of payload bytes}

\item{info}{\code{sensorColumns} information from a \code{SENSOR_SCHEMA}
object}

\item{id}{integer. The \code{id} information from a \code{SENSOR_SCHEMA}
object}
}
\description{
Parse SENSOR_DATA packet in c++
}
\keyword{internal}
