% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PARAMETERS_payload.R
\name{process_parameters}
\alias{process_parameters}
\title{Second-level function for parsing PARAMETERS packets}
\usage{
process_parameters(payload_chunk, tz)
}
\arguments{
\item{tz}{character. The timezone to use}
}
\description{
This is a helper function for \code{\link{parse_packet_set.PARAMETERS}}. It
dispatches to third-level functions (e.g. \code{\link{par_battery_state}})
that correspond to each packet address and identifier.
}
\keyword{internal}
