% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_agd.R
\name{read_agd}
\alias{read_agd}
\title{Read an ActiGraph agd file}
\usage{
read_agd(file, tz = "UTC", return = c("data", "settings", "both"))
}
\arguments{
\item{file}{character. Path to the agd file}

\item{tz}{character. The desired timezone}

\item{return}{character. The desired elements to return (see details)}
}
\value{
the desired output
}
\description{
Read an ActiGraph agd file
}
\details{
The agd is made up of a \code{settings} element and a \code{data}
  element. The former is likely only necessary for debugging, and so the
  default is to return only the \code{data} element. Both \code{settings} and
  \code{data} are in data frame format. If both are returned, they are given
  as a list.
}
\seealso{
\url{ActiGraph documentation}{https://github.com/actigraph/ActiLifeManual/blob/master/docs/appendix.rst}
}
