% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_setup.R
\name{unzip_gt3x}
\alias{unzip_gt3x}
\alias{check_gt3x_components}
\alias{read_gt3x_setup}
\title{Prepare a gt3x file for parsing}
\usage{
unzip_gt3x(file)

check_gt3x_components(file, verbose = FALSE)

read_gt3x_setup(file, verbose = FALSE)
}
\arguments{
\item{file}{character. Path to the file}

\item{verbose}{logical. Print updates to console?}
}
\description{
Prepare a gt3x file for parsing
}
\keyword{internal}
