% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_utils.R
\name{tick_to_posix}
\alias{tick_to_posix}
\title{Convert a tick value to a timestamp}
\usage{
tick_to_posix(x, tz = "UTC", ...)
}
\arguments{
\item{x}{the tick value}

\item{tz}{character. The timezone to use}

\item{...}{further arguments passed to \code{as.POSIXct}}
}
\description{
Convert a tick value to a timestamp
}
\keyword{internal}
