% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gt3x_parse_log_bin_utils.R
\name{validate_include}
\alias{validate_include}
\title{Validate the values provided for the \code{input} argument in
\code{\link{read_gt3x}}}
\usage{
validate_include(include, verbose = FALSE)
}
\arguments{
\item{include}{character. The PACKET types to parse}

\item{verbose}{logical. Print updates to console?}
}
\description{
Validate the values provided for the \code{input} argument in
\code{\link{read_gt3x}}
}
\keyword{internal}
