% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_log_bin_utils.R
\name{validate_include}
\alias{validate_include}
\title{Validate the values provided for the \code{input} argument in
\code{\link{read_gt3x}}}
\usage{
validate_include(include, verbose = FALSE, choices = c("METADATA",
  "PARAMETERS", "SENSOR_SCHEMA", "BATTERY", "EVENT", "TAG", "ACTIVITY",
  "HEART_RATE_BPM", "HEART_RATE_ANT", "HEART_RATE_BLE", "LUX", "CAPSENSE",
  "EPOCH", "EPOCH2", "EPOCH3", "EPOCH4", "ACTIVITY2", "SENSOR_DATA"))
}
\arguments{
\item{include}{character. The PACKET types to parse}

\item{verbose}{logical. Print updates to console?}

\item{choices}{character. The packet types to choose from.}
}
\description{
Validate the values provided for the \code{input} argument in
\code{\link{read_gt3x}}
}
\keyword{internal}
